pyproject.toml
[tool.poetry]
name = "databrics-demo"
version = "0.1.0"
description = ""
authors = ["ranga nalluri <nalluri333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.7"
mlflow = {extras = ["databricks"], version = "^2.18.0"}
scikit-learn = "^1.5.2"
databricks-sql-connector = "^3.6.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
